SS: rename sites files to improve readability
