+++ sites/texas_tribune.py
@@ -9,10 +9,9 @@ from bs4 import BeautifulSoup
 from requests.models import Response
 
 from sites.config.config import ConfigCF, ConfigPop, ScrapeConfig, SiteConfig, TrainParamsCF
-from sites.helpers.google_tag_manager import GOOGLE_TAG_MANAGER_RAW_FIELDS, transform_data_google_tag_manager
-from sites.helpers.scrape_error import ArticleScrapingError, ScrapeFailure, ArticleBatchScrapingError
-from sites.helpers.safe_get import safe_get
-from sites.helpers.validate import validate_response, validate_status_code
+from sites.helpers.gtm import GOOGLE_TAG_MANAGER_RAW_FIELDS, transform_data_google_tag_manager
+from sites.helpers.requests import ArticleScrapingError, ScrapeFailure, ArticleBatchScrapingError
+from sites.helpers.requests import safe_get, validate_response, validate_status_code
 from sites.templates.site import Site
 from sites.config.strategy import Strategy
