+++ tests/sites/test_helpers.py
@@ -1,7 +1,7 @@
 import pytest
 from requests.models import Response
 
-from sites.helpers.validate import validate_response, validate_status_code
+from sites.helpers.requests import validate_response, validate_status_code
 
 
 def _create_response(status_code: int) -> Response:
