+++ tests/test_scrape_metadata.py
@@ -6,7 +6,7 @@ import pandas as pd
 from db.mappings.article import Article
 from db.mappings.path import Path
 from job.steps.collaborative_filtering.scrape_metadata import scrape_upload_metadata
-from sites.helpers.scrape_error import ArticleScrapingError, ScrapeFailure
+from sites.helpers.requests import ArticleScrapingError, ScrapeFailure
 from sites.templates.site import Site
 from sites.sites import Sites
 from tests.base import BaseTest
